# Add additional metrics and tests to FVM bench