1. Added 2 new bench tests.
2. Added reporting computation used and interaction used to all bench tests
3. The output of benchstat will now be longer, so I changed it to be collapsed following this guide: https://gist.github.com/pierrejoubert73/902cc94d79424356a8d20be2b382e1ab

This should give us some more metrics to look at, and will make it easier to detect and fix high ledger interaction usage.

results will looks similar to:
```
name                                                                old time/op       new time/op       delta
RuntimeNFTBatchTransfer-8                                                86.4ms ±11%       80.8ms ± 4%     ~     (p=0.095 n=5+5)
RuntimeTransaction/reference_tx-8                                        25.0ms ±11%       23.6ms ±14%     ~     (p=0.151 n=5+5)
RuntimeTransaction/convert_int_to_string-8                               27.3ms ±20%       25.3ms ±15%     ~     (p=0.841 n=5+5)
RuntimeTransaction/convert_int_to_string_and_concatenate_it-8            26.8ms ±22%       25.8ms ± 8%     ~     (p=1.000 n=5+5)
RuntimeTransaction/get_signer_address-8                                  28.8ms ±42%       25.0ms ±21%     ~     (p=0.421 n=5+5)
RuntimeTransaction/get_public_account-8                                  29.9ms ±30%       26.8ms ±16%     ~     (p=0.222 n=5+5)
RuntimeTransaction/get_account_and_get_balance-8                          422ms ±38%        401ms ±31%     ~     (p=0.841 n=5+5)
RuntimeTransaction/get_account_and_get_available_balance-8                333ms ± 7%        304ms ±18%     ~     (p=0.151 n=5+5)
RuntimeTransaction/get_account_and_get_storage_used-8                    41.7ms ±36%       34.2ms ±10%     ~     (p=0.095 n=5+5)
RuntimeTransaction/get_account_and_get_storage_capacity-8                 360ms ±47%        290ms ±44%     ~     (p=0.151 n=5+5)
RuntimeTransaction/get_signer_vault-8                                    32.7ms ±12%       29.3ms ± 6%     ~     (p=0.056 n=5+5)
RuntimeTransaction/get_signer_receiver-8                                 48.7ms ±58%       37.2ms ± 9%  -23.45%  (p=0.032 n=5+5)
RuntimeTransaction/transfer_tokens-8                                      147ms ±13%        131ms ± 6%     ~     (p=0.151 n=5+5)
RuntimeTransaction/load_and_save_empty_string_on_signers_address-8       29.7ms ±19%       28.5ms ±13%     ~     (p=0.548 n=5+5)
RuntimeTransaction/load_and_save_long_string_on_signers_address-8        31.5ms ±11%       31.2ms ±30%     ~     (p=0.690 n=5+5)
RuntimeTransaction/create_new_account-8                                   777ms ±10%        732ms ± 5%     ~     (p=0.111 n=5+4)
RuntimeTransaction/call_empty_contract_function-8                        26.7ms ± 5%       26.9ms ± 9%     ~     (p=1.000 n=5+5)
RuntimeTransaction/emit_event-8                                          32.2ms ± 5%       31.6ms ± 6%     ~     (p=0.841 n=5+5)
RuntimeTransaction/borrow_array_from_storage-8                           93.8ms ±14%       90.7ms ± 6%     ~     (p=0.841 n=5+5)
RuntimeTransaction/copy_array_from_storage-8                             84.2ms ± 2%       88.6ms ±16%     ~     (p=0.310 n=5+5)

name                                                                old computation   new computation   delta
RuntimeTransaction/reference_tx-8                                           202 ± 0%          202 ± 0%     ~     (all equal)
RuntimeTransaction/convert_int_to_string-8                                  402 ± 0%          402 ± 0%     ~     (all equal)
RuntimeTransaction/convert_int_to_string_and_concatenate_it-8               502 ± 0%          502 ± 0%     ~     (all equal)
RuntimeTransaction/get_signer_address-8                                     302 ± 0%          302 ± 0%     ~     (all equal)
RuntimeTransaction/get_public_account-8                                     402 ± 0%          402 ± 0%     ~     (all equal)
RuntimeTransaction/get_account_and_get_balance-8                            402 ± 0%          402 ± 0%     ~     (all equal)
RuntimeTransaction/get_account_and_get_available_balance-8                  402 ± 0%          402 ± 0%     ~     (all equal)
RuntimeTransaction/get_account_and_get_storage_used-8                       402 ± 0%          402 ± 0%     ~     (all equal)
RuntimeTransaction/get_account_and_get_storage_capacity-8                   402 ± 0%          402 ± 0%     ~     (all equal)
RuntimeTransaction/get_signer_vault-8                                       402 ± 0%          402 ± 0%     ~     (all equal)
RuntimeTransaction/get_signer_receiver-8                                    602 ± 0%          602 ± 0%     ~     (all equal)
RuntimeTransaction/transfer_tokens-8                                      3.50k ± 0%        3.50k ± 0%     ~     (all equal)
RuntimeTransaction/load_and_save_empty_string_on_signers_address-8          602 ± 0%          602 ± 0%     ~     (all equal)
RuntimeTransaction/load_and_save_long_string_on_signers_address-8           602 ± 0%          602 ± 0%     ~     (all equal)
RuntimeTransaction/create_new_account-8                                     202 ± 0%          202 ± 0%     ~     (all equal)
RuntimeTransaction/call_empty_contract_function-8                           402 ± 0%          402 ± 0%     ~     (all equal)
RuntimeTransaction/emit_event-8                                             602 ± 0%          602 ± 0%     ~     (all equal)
RuntimeTransaction/borrow_array_from_storage-8                            2.60k ± 0%        2.60k ± 0%     ~     (all equal)
RuntimeTransaction/copy_array_from_storage-8                              2.60k ± 0%        2.60k ± 0%     ~     (all equal)

name                                                                old interactions  new interactions  delta
RuntimeTransaction/reference_tx-8                                          354k ± 0%         354k ± 0%     ~     (all equal)
RuntimeTransaction/convert_int_to_string-8                                 354k ± 0%         354k ± 0%     ~     (all equal)
RuntimeTransaction/convert_int_to_string_and_concatenate_it-8              354k ± 0%         354k ± 0%     ~     (all equal)
RuntimeTransaction/get_signer_address-8                                    354k ± 0%         354k ± 0%     ~     (all equal)
RuntimeTransaction/get_public_account-8                                    354k ± 0%         354k ± 0%     ~     (all equal)
RuntimeTransaction/get_account_and_get_balance-8                          15.9M ± 0%        15.9M ± 0%     ~     (all equal)
RuntimeTransaction/get_account_and_get_available_balance-8                5.45M ± 0%        5.45M ± 0%     ~     (all equal)
RuntimeTransaction/get_account_and_get_storage_used-8                      357k ± 0%         357k ± 0%     ~     (all equal)
RuntimeTransaction/get_account_and_get_storage_capacity-8                 5.45M ± 0%        5.45M ± 0%     ~     (all equal)
RuntimeTransaction/get_signer_vault-8                                      354k ± 0%         354k ± 0%     ~     (all equal)
RuntimeTransaction/get_signer_receiver-8                                   354k ± 0%         354k ± 0%     ~     (all equal)
RuntimeTransaction/transfer_tokens-8                                       354k ± 0%         354k ± 0%     ~     (all equal)
RuntimeTransaction/load_and_save_empty_string_on_signers_address-8         354k ± 0%         354k ± 0%     ~     (all equal)
RuntimeTransaction/load_and_save_long_string_on_signers_address-8          357k ± 0%         357k ± 0%     ~     (all equal)
RuntimeTransaction/create_new_account-8                                   10.7M ± 0%        10.7M ± 0%     ~     (all equal)
RuntimeTransaction/call_empty_contract_function-8                          354k ± 0%         354k ± 0%     ~     (all equal)
RuntimeTransaction/emit_event-8                                            354k ± 0%         354k ± 0%     ~     (all equal)
RuntimeTransaction/borrow_array_from_storage-8                             359k ± 0%         359k ± 0%     ~     (all equal)
RuntimeTransaction/copy_array_from_storage-8                               359k ± 0%         359k ± 0%     ~     (all equal)
```